pyproject.toml
[tool.poetry]
name = "fsrbeschluesse"
version = "0.1.0"
description = ""
authors = ["Jonas Gaffke <jonas@jonasga.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
sqlmodel = "^0.0.8"
PyMySQL = "^1.0.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
